requirements.txt
Click==7.0
Flask==1.1.1
get==2019.4.13
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
numpy==1.17.0
pandas==0.25.0
post==2019.4.13
public==2019.4.13
pymongo==3.8.0
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
query-string==2019.4.13
request==2019.4.13
six==1.12.0
Werkzeug==0.15.5
